refactor(ui): update padding variant and restructure CardVariant for improved consistency in ResourceStatsCard components
